+++ Engine/src/common/Utils.hpp
@@ -13,13 +13,13 @@ enum ShapeType
 
 struct Color
 {
-    int r;
-    int g;
-    int b;
-    int a;
+    uint8_t r;
+    uint8_t g;
+    uint8_t b;
+    uint8_t a;
 
     // Constructor to initialize color values using RGB values
-    Color(int red, int green, int blue, int alpha = 255)
+    Color(uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha = 255)
         : r(clamp(red)), g(clamp(green)), b(clamp(blue)), a(clamp(alpha)) {}
 
     static const Color RED;
